requirements.txt
arrow==0.12.1
blessed==1.15.0
boto==2.49.0
boto3==1.9.24
botocore==1.12.24
dj-database-url==0.5.0
Django==1.11.16
django-elasticsearch-dsl==0.5.0
django-environ==0.4.5
django-picklefield==1.1.0
django-q==1.0.1
django-ses==0.8.7
django-storages==1.7.1
django-versatileimagefield==1.10
docutils==0.14
elasticsearch==5.5.3
elasticsearch-dsl==5.4.0
future==0.16.0
gunicorn==19.9.0
ipaddress==1.0.22
jmespath==0.9.3
mysqlclient==1.3.13
Pillow==5.3.0
psycopg2==2.7.5
python-dateutil==2.7.3
pytz==2018.5
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
wcwidth==0.1.7
whitenoise==4.1
